feat: Add version checking
Since the recommended configs may change over time, it's important
that users stay up to date. This commit adds a facility for
checking for updates and notifying the user if a newer version
is available.

For now we're checking on every command invocation, but we can
cache this in the future if it becomes problematic.